src/backend-python/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
mysql-connector-python==8.0.33
PyJWT==2.8.0
Werkzeug==2.3.7
